Refactor local classifier training and evaluation pipeline
- Changed HMCLocalModel to use nn.ModuleDict instead of nn.ModuleList for levels.
- Created a new main training script for local classifiers, consolidating training, validation, and testing steps.
- Implemented separate validation and testing functions to improve code organization.
- Removed the old train_local.py file and integrated its functionality into the new structure.
- Updated logging to provide more informative outputs during training and evaluation.
- Ensured compatibility with hyperparameter optimization and data preprocessing steps.
