+++ src/hmc/train/utils.py
@@ -1,6 +1,6 @@
 import json
 from datetime import datetime
-
+import logging
 import numpy as np
 
 
@@ -28,12 +28,12 @@ def local_to_global_predictions(local_labels, local_nodes_idx, nodes_idx):
         level: {v: k for k, v in local_nodes_idx[level].items()}
         for level in sorted_levels
     }
-    # print(f"Local nodes idx: {local_nodes_idx}")
-    # print(f"Local nodes reverse: {local_nodes_reverse}")
+    # logging.info(f"Local nodes idx: {local_nodes_idx}")
+    # logging.info(f"Local nodes reverse: {local_nodes_reverse}")
 
-    print(f"Exemplos: {n_samples}")
-    # print(f"Shape local_preds: {len(local_labels)}")
-    # print(f"Local nodes idx: {local_nodes_reverse}")
+    logging.info(f"Exemplos: {n_samples}")
+    # logging.info(f"Shape local_preds: {len(local_labels)}")
+    # logging.info(f"Local nodes idx: {local_nodes_reverse}")
 
     # Etapa 1: montar node_names ativados por exemplo
     activated_nodes_by_example = [[] for _ in range(n_samples)]
@@ -49,18 +49,18 @@ def local_to_global_predictions(local_labels, local_nodes_idx, nodes_idx):
                 if node_name:
                     activated_nodes_by_example[idx_example].append(node_name)
                 else:
-                    print(
+                    logging.info(
                         f"[WARN] Índice local {local_idx} não encontrado no nível {level}"
                     )
 
-    # print(f"Node names ativados por exemplo: {activated_nodes_by_example[0]}")
+    # logging.info(f"Node names ativados por exemplo: {activated_nodes_by_example[0]}")
     global_indices = []
     for node in activated_nodes_by_example[0]:
-        # print(f"Node names ativados: {node}")
+        # logging.info(f"Node names ativados: {node}")
         if "/" in node:
             node = node.replace("/", ".")
         global_indices.append(nodes_idx.get(node))
-    print(global_indices)
+    logging.info(global_indices)
     # Etapa 2: converter node_names para índices globais
     for idx_example, node_names in enumerate(activated_nodes_by_example):
         for node_name in node_names:
@@ -69,43 +69,39 @@ def local_to_global_predictions(local_labels, local_nodes_idx, nodes_idx):
                 global_idx = nodes_idx[key]
                 global_preds[idx_example][global_idx] = 1
             else:
-                print(f"[WARN] Node '{key}' não encontrado em nodes_idx")
+                logging.info(f"[WARN] Node '{key}' não encontrado em nodes_idx")
 
     return global_preds
 
 
-def show_local_losses(local_losses, set="Train"):
-    formated_string = ""
+def show_local_losses(local_losses, dataset="Train"):
+    formatted_string = ""
     for level, local_loss in enumerate(local_losses):
         if local_loss is not None and local_loss != 0.0:
-            if level == 0:
-                formated_string += f"{set} Loss {local_loss:.4f} for level {level} // "
-            else:
-                formated_string += (
-                    f" + {set} Loss {local_loss:.4f} for level {level} // "
-                )
+            formatted_string += "%s Loss %s for level %s // " % (
+                dataset,
+                local_loss,
+                level,
+            )
 
-    print(formated_string)
-
-
-def show_global_loss(global_loss, set="Train"):
-    print(f"Global average loss {set} Loss: {global_loss:.4f}")
+    logging.info(formatted_string)
 
 
-def show_local_precision(local_precisions, set="Train"):
-    formated_string = ""
-    for level, local_precision in enumerate(local_precisions):
-        if local_precision is not None and local_precision != 0.0:
-            if level == 0:
-                formated_string += (
-                    f"{set} Precision {local_precision:.4f} for level {level} // "
-                )
-            else:
-                formated_string += (
-                    f" + {set} Precision {local_precision:.4f} for level {level} // "
-                )
+def show_global_loss(global_loss, dataset="Train"):
+    logging.info("Global average loss %s Loss: %s", dataset, global_loss)
 
-    print(formated_string)
+
+def show_local_score(local_scores, dataset="Train"):
+    formatted_string = ""
+    for level, local_score in enumerate(local_scores):
+        if local_score is not None and local_score != 0.0:
+            formatted_string += "%s Score %s for level %s // " % (
+                dataset,
+                local_score,
+                level,
+            )
+
+    logging.info(formatted_string)
 
 
 def save_dict_to_json(dictionary, file_path):
